moved package `fynemd` out
