+++ fynemd/ScaledText.go
